requirements.txt
django
djangorestframework
django-cors-headers
pillow
djongo
pymongo
tensorflow
numpy
pandas
opencv-python-headless
pytesseract
djangorestframework-simplejwt
gunicorn
whitenoise
scikit-learn
python-decouple
requests
ipython
django-debug-toolbar
boto3
django-storages
tensorflow
numpy
pandas
opencv-python-headless
pytesseract
scikit-learn
matplotlib
seaborn
